Create start.bat
This is for the first time i do somting at github so don't blame me if  i do wrong. For the noobs of us a easy start.bat, new versions in the same folder will picked up and run. 

This weekend i try to ask some questions at the begin of command promp with username / pass and start coordinate and write this in the config file.